refactor: rename pkg/optimizer to pkg/advisor (#243)
